[Execute] Revisit executeQuickInput file
Let's revisit executeQuickInput file.

ONE-vscode-DCO-1.0-Signed-off-by: Yongseop Kim <yons.kim@samsung.com>
