+++ package.json
@@ -283,7 +283,7 @@
       "commandPalette": [
         {
           "command": "one.backend.infer",
-          "when": "false"
+          "when": "true"
         },
         {
           "command": "one.device.register",
