+++ src/Execute/executeQuickInput.ts
@@ -19,152 +19,62 @@ import {appendFileSync} from 'fs';
 import * as vscode from 'vscode';
 
 import {Backend} from '../Backend/API';
-import {globalBackendMap} from '../Backend/Backend';
 import {Logger} from '../Utils/Logger';
-import {MultiStepInput} from '../Utils/MultiStepInput';
 
-export async function runInferenceQuickInput(context: vscode.ExtensionContext) {
-  const logTag = 'runInferenceQuickInput';
+import {collectInputs} from './InferenceQuickInput';
 
-  interface State {
-    backend: Backend;
-    modelPath: vscode.Uri;
-    inputSpec: string;
-    selectedItem: vscode.QuickPickItem;
-    error: string|undefined;
-  }
 
-  async function collectInputs() {
-    const state = {error: undefined} as Partial<State>;
-    await MultiStepInput.run(input => pickBackend(input, state));
-    return state as State;
-  }
+const logTag = 'runInferenceQuickInput';
 
-  async function pickBackend(input: MultiStepInput, state: Partial<State>) {
-    const backends: vscode.QuickPickItem[] = Object.keys(globalBackendMap).map(label => ({label}));
-    state.selectedItem = await input.showQuickPick({
-      title: 'Choose Executor Toolchain',
-      step: 1,
-      totalSteps: 3,
-      placeholder: 'Select a Backend',
-      items: backends,
-      shouldResume: shouldResume
-    });
-    state.backend = globalBackendMap[state.selectedItem.label];
-    return (input: MultiStepInput) => selectInputModel(input, state);
-  }
+async function runInference(
+    backend: Backend, modelPath: vscode.Uri, inputSpec: string): Promise<void> {
+  let cmd = backend.executor() ?.runInference(modelPath.path, ['--input-spec', inputSpec]);
+  let outFileName: string = `${modelPath.path}.infer.log`;
 
-  // TODO: Use quickPick window with fast grep child process
-  async function selectInputModel(input: MultiStepInput, state: Partial<State>) {
-    if (state.backend === undefined) {
-      state.error = 'Backend to infer is not chosen. Please check once again.';
-      return undefined;
-    }
-    if (state.backend.executor() === undefined) {
-      state.error = 'Backend executor is not set yet. Please check once again.';
-      return undefined;
-    }
-
-    let backendName: string = state.backend.name();
-    const executor = state.backend.executor();
-    let filter: {[name: string]: string[]} = {};
-    // List files which are filtered with executable extensions
-    if (executor) {
-      filter = {backendName: executor.getExecutableExt()};
-    }
-
-    const fileUri = await vscode.window.showOpenDialog({
-      title: `Select Model to Infer (2/3)`,
-      canSelectMany: false,
-      openLabel: 'Select Model to Infer',
-      filters: filter
-    });
-    if (fileUri && fileUri[0]) {
-      state.modelPath = fileUri[0];
-    } else {
-      Logger.warn(logTag, 'No model has been selected');
-      state.error = 'No model has been selected. Please check once again.';
-      state.modelPath = undefined;
-      return undefined;
-    }
-
-    return (input: MultiStepInput) => selectInputSpec(input, state);
-  }
-
-  // TODO: enable the backend-driven option steps by backend extension
-  async function selectInputSpec(input: MultiStepInput, state: Partial<State>) {
-    const inputSpecKeys: vscode.QuickPickItem[] =
-        ['any', 'non-zero', 'positive'].map(label => ({label}));
-
-    state.selectedItem = await input.showQuickPick({
-      title: 'Enter Backend Specific Options',
-      step: 3,
-      totalSteps: 3,
-      placeholder: 'Select Random Input Spec',
-      items: inputSpecKeys,
-      shouldResume: shouldResume
-    });
-    state.inputSpec = state.selectedItem.label;
-  }
-
-  async function runInference(state: State) {
-    if (state.error !== undefined) {
-      vscode.window.showErrorMessage(state.error);
-      return;
-    }
-    let cmd =
-        state.backend.executor() ?.runInference(
-                                      state.modelPath.path, ['--input-spec', state.inputSpec]);
-    let outFileName: string = `${state.modelPath.path}.infer.log`;
-
-    await vscode.window
-        .withProgress(
-            {
-              location: vscode.ProgressLocation.Notification,
-              title: 'Inference Running!',
-              cancellable: true
-            },
-            (progress, token) => {
-              token.onCancellationRequested(() => {
-                Logger.info(logTag, 'User canceled the log running operation');
-              });
-              const p = new Promise((resolve, reject) => {
-        exec(cmd?.str() + ' > ' + outFileName, (error, stdout, stderr) => {
-          if (error) {
-            return reject(error.message);
-          }
-          // Some of warnings are treated as error.
-          // TODO: handle the stderr
-          // if (stderr) return reject(stderr);
-          else {
-            return resolve(stdout);
-          }
-        });
-              });
-              return p;
-            })
-        .then(
-            () => {
-              vscode.window.showInformationMessage(
-                  `Inference succeeded! You can find the log at ${outFileName}`, {title: 'OK'});
-            },
-            (error: string) => {
-              appendFileSync(outFileName, error);
-              // TODO: Find if the message includes new line character.
-              vscode.window.showErrorMessage(
-                  'Exception Occurred! You can find the log at ' + outFileName, {title: 'Close'});
+  await vscode.window
+      .withProgress(
+          {
+            location: vscode.ProgressLocation.Notification,
+            title: 'Inference Running!',
+            cancellable: true
+          },
+          (progress, token) => {
+            token.onCancellationRequested(() => {
+              Logger.info(logTag, 'User canceled the log running operation');
             });
-  }
+            const p = new Promise((resolve, reject) => {
+      exec(cmd?.str() + ' > ' + outFileName, (error, stdout, stderr) => {
+        if (error) {
+          return reject(error.message);
+        }
+        // Some of warnings are treated as error.
+        // TODO: handle the stderr
+        // if (stderr) return reject(stderr);
+        else {
+          return resolve(stdout);
+        }
+      });
+            });
+            return p;
+          })
+      .then(
+          () => {
+            vscode.window.showInformationMessage(
+                `Inference succeeded! You can find the log at ${outFileName}`, {title: 'OK'});
+          },
+          (error: string) => {
+            appendFileSync(outFileName, error);
+            // TODO: Find if the message includes new line character.
+            vscode.window.showErrorMessage(
+                'Exception Occurred! You can find the log at ' + outFileName, {title: 'Close'});
+          });
+}
 
-
-  function shouldResume() {
-    // Could show a notification with the option to resume.
-    return new Promise<boolean>(
-        (resolve, reject) => {
-            // noop
-        });
-  }
-
+export async function runInferenceQuickInput(context: vscode.ExtensionContext): Promise<void> {
   const state = await collectInputs();
-  await runInference(state);
+  if (state.error !== undefined) {
+    vscode.window.showErrorMessage(state.error);
+    return;
+  }
+  await runInference(state.backend, state.modelPath, state.inputSpec);
 }
